Split server-attendee-form tests into a folder and drop its max-lines exemption
The unified-attendee-form server test file had grown to 1,913 lines and
was grandfathered into biome's noExcessiveLinesPerFile exemption list.
Split it into test/lib/server-attendee-form/ — grouped by area (new form,
create & line edits, no-quantity, edit-page panels, overbooking & mixed
timing, error paths, custom questions, and status/balance/tabs) — so every
file sits under 400 lines, then removed the file's exemption entry.

Bringing the file back under jscpd's per-file scan cap surfaced repeated
attendee-form setup that had been passing silently inside the monolith. The
shared pieces are now a folder-local helpers.ts: submitNewAttendeeForm,
expectAttendeeLineCount, everydayDailyListing, tomorrowInTz,
attendeeWithNoBookings, and expectMixedStandardAndDailyLines. All original
tests carry over unchanged and pass (107 across the folder); jscpd is back
to 0% duplication.
