+++ biome.json
@@ -171,7 +171,6 @@
         "src/ui/templates/fields.ts",
         "src/ui/templates/public/reservations.tsx",
         "test/e2e/accounting.test.ts",
-        "test/lib/server-attendee-form.test.ts",
         "test/lib/server-attendees.test.ts",
         "test/lib/server-listing-parents.test.ts",
         "test/lib/server-parents-booking.test.ts",
